feat(WalletButton): Connecting button state
fix(WalletButton): Margin and mobile fixes
