+++ src/components/common/WalletButton/styles.module.scss
@@ -1,61 +1,61 @@
 @import "../../../assets/styles/_mixins.scss";
 
 .balance_panel {
-  @include flexbox(center, center);
+    @include flexbox(center, center);
+    width: fit-content;
+    max-width: 210px;
+    height: 33px !important;
+    background: var(--color-blue);
+    border: 1px solid var(--color-blue);
+    border-radius: 10px;
+    color: var(--color-white);
+    font-family: $default-font;
+    font-style: normal;
+    padding: 8px 16px;
+    cursor: pointer;
 
-  max-width: 210px;
-  min-width: 0;
-  height: 33px !important;
-  background: var(--color-blue);
-  border: 1px solid var(--color-blue);
-  border-radius: 10px;
-  color: var(--color-white);
-  font-family: $default-font;
-  font-style: normal;
-  padding: 0;
-  cursor: pointer;
-
-  @include tablet {
-    margin-top: 10px;
-    width: 100%;
-    max-width: 100%;
-  }
+    @include tablet {
+        width: 100%;
+        max-width: 100%;
+    }
 }
 
 .icon_wrap {
-  padding: 8.25px 19px 9px 0;
-  margin: 0 0 0 auto;
+    margin: 0 0 0 auto;
 }
 
-.icon {
-  @include flexbox(center, center);
 
-  color: var(--color-white);
-  width: 12px;
-  height: 15.75px;
+.icon {
+    @include flexbox(center, center);
+    margin-left: 8px;
+    color: var(--color-white);
+    width: 12px;
+    height: 15.75px;
 }
 
 .balance {
-  @include flexbox(center, center);
+    @include flexbox(center, center);
+    vertical-align: middle;
+    font-weight: bold;
+    font-size: 14px;
+    letter-spacing: -0.01em;
+    font-weight: 700px;
+    white-space: nowrap;
+    margin-right: 8px;
+}
 
-  padding: 8px 0 8px 16px;
-  font-weight: bold;
-  font-size: 14px;
-  line-height: 17px;
-  letter-spacing: -0.01em;
-  font-weight: 700px;
-  white-space: nowrap;
-  margin: 0 10px 0 0;
+.icon-wrap:empty, .balance:empty {
+    display: none;
 }
 .address {
-  font-weight: normal;
-  font-size: 12px;
-  line-height: 17px;
-  letter-spacing: -0.01em;
-  padding: 8px 0 8px 0;
-  margin: 0 7px 0 0;
+    @include flexbox(center, center);
+    font-weight: normal;
+    font-size: 12px;
+    letter-spacing: -0.01em;
 
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
+    vertical-align: middle;
+
+    text-overflow: ellipsis;
+    overflow: hidden;
+    white-space: nowrap;
 }
