+++ src/store/main/sagas/mainCheckSaga.ts
@@ -4,7 +4,6 @@ import { ModalType } from 'store/modal/types';
 import Web3 from 'web3';
 import Web3Modal from 'web3modal';
 import WalletConnectProvider from '@walletconnect/web3-provider';
-
 import {
   mainCheck,
   mainGetData,
